backend/requirements.txt
fastapi>=0.109.0,<1
uvicorn[standard]>=0.27.0,<1
sqlalchemy>=2.0.0,<3
qrcode[pil]>=7.4.2,<8
python-jose[cryptography]>=3.3.0,<4
bcrypt>=4.0.0,<5
email-validator>=2.0.0,<3
